Unify setup script exampel
